src: rename legacy trace event headers
