+++ src/inspector_agent.cc
@@ -14,7 +14,9 @@
 #include "inspector/runtime_agent.h"
 #include "inspector/storage_agent.h"
 #include "inspector/target_agent.h"
+#ifndef V8_USE_PERFETTO
 #include "inspector/tracing_agent.h"
+#endif  // V8_USE_PERFETTO
 #include "inspector/worker_agent.h"
 #include "inspector/worker_inspector.h"
 #include "inspector_io.h"
@@ -238,9 +240,11 @@ class ChannelImpl final : public v8_inspector::V8Inspector::Channel,
                                   StringView(),
                                   V8Inspector::ClientTrustLevel::kFullyTrusted);
     node_dispatcher_ = std::make_unique<UberDispatcher>(this);
+#ifndef V8_USE_PERFETTO
     tracing_agent_ =
         std::make_unique<protocol::TracingAgent>(env, main_thread_);
     tracing_agent_->Wire(node_dispatcher_.get());
+#endif  // V8_USE_PERFETTO
     if (worker_manager) {
       worker_agent_ = std::make_unique<protocol::WorkerAgent>(worker_manager);
       worker_agent_->Wire(node_dispatcher_.get());
@@ -274,8 +278,10 @@ class ChannelImpl final : public v8_inspector::V8Inspector::Channel,
   }
 
   ~ChannelImpl() override {
+#ifndef V8_USE_PERFETTO
     tracing_agent_->disable();
     tracing_agent_.reset();  // Dispose before the dispatchers
+#endif                       // V8_USE_PERFETTO
     if (worker_agent_) {
       worker_agent_->disable();
       worker_agent_.reset();  // Dispose before the dispatchers
@@ -436,7 +442,9 @@ class ChannelImpl final : public v8_inspector::V8Inspector::Channel,
   }
 
   std::unique_ptr<protocol::RuntimeAgent> runtime_agent_;
+#ifndef V8_USE_PERFETTO
   std::unique_ptr<protocol::TracingAgent> tracing_agent_;
+#endif
   std::unique_ptr<protocol::WorkerAgent> worker_agent_;
   std::shared_ptr<protocol::TargetAgent> target_agent_;
   std::unique_ptr<NetworkInspector> network_inspector_;
